Adding 2.6.32 folder.
This folder contains the complete NDAS driver that was in formerly
available in part at code.ximeta.com. It is available for
community development. Some instructions are in the readme files
and an overview is possible with make help.

Please post questions and patches to the ticket server at
ndas4linux.iocellnetworks.com so others can participate and help.