3.7-rc5 removed.
The version was transferred to linuxdriverproject for preliminary
consideration, before entering to staging development.
3.7-rc5 removed.

The version was transferred to linuxdriverproject for preliminary
consideration, before entering to staging development.
3.7-rc5 removed.

The version was transferred to linuxdriverproject for preliminary
consideration, before entering to staging development.
3.7-rc5 removed.

The version was transferred to linuxdriverproject for preliminary
consideration, before entering to staging development.
3.7-rc5 removed.

The version was transferred to linuxdriverproject for preliminary
consideration, before entering to staging development.
3.7-rc5 removed.

The version was transferred to linuxdriverproject for preliminary
consideration, before entering to staging development.
3.7-rc5 removed.

The version was transferred to linuxdriverproject for preliminary
consideration, before entering to staging development.